# public/openapi.yaml
info:
  description: |
    Purpie is an open source social media solution focusing on video oriented communication and content sharing.
    
    The Purpie endpoints are offered to our users to facilitate the use of our technology and allow seamless integrations with their applications.
    For this documentation, you can use the **Api Key** and **Api Secret** which generated from [Purpie](https://purpie.io/) settings page to test the services.

    # Introduction
    Purpie orchestrates various media services and solutions for online meetings, live streamings, on demand video delivery and many more.
    With an agile and hierarchical way of user interaction and content sharing possibilities, Purpie proposes a robust and flexible solution for
    various video communication and distribution needs.

    # Cross-Origin Resource Sharing
    This API features Cross-Origin Resource Sharing (CORS) implemented.
    And that allows cross-domain communication from the browser.
    All responses have a wildcard same-origin which makes them completely public and accessible to everyone, including any code on any site.

    # Authentication
    To make authorized calls on Purpie API operations, you need to generate an API key and secret in the settings page after login to [Purpie](https://purpie.io/).
    
    After generating API credentials, you need to note your `api secret`. You can access this field once after creating. If you lost or forgot your api
...
    
...
    header:
    
...

...
  x-logo:
...
servers:
...
openapi: 3.0.0
paths:
...
  /v1/zone/create:
    post:
      operationId: ZoneController_createZone
      summary: Create Zone
      description: Current authenticated user creates a new zone. The id of the user zone is returned. User must have "canCreateZone" permission.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateZoneDto'
      responses:
        '201':
          description: New zone created.
          content:
            application/json:
              schema:
                type: integer
        '400':
          description: Error thrown when new zone subdomain is not unique.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 403
                  message:
                    type: string
                    example: The zone with requested subdomain already exist.
                  error:
                    type: string
                    example: ZONE_SUBDOMAIN_ALREADY_EXIST
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '403':
          description: Error thrown when current user membership zone creation limit is full
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 403
                  message:
                    type: string
                    example: Your channel zone operation failed due to insufficient membership.
                  error:
                    type: string
                    example: INSUFFICIENT_MEMBERSHIP
      tags:
        - Zone
  /v1/zone/update/{zoneId}:
    put:
      operationId: ZoneController_editZone
      summary: Create Zone
      description: User updates a zone belonging to zoneId with given parameters. User must have the "canEdit" permission on zone.
      parameters:
        - name: zoneId
          required: true
          in: path
          description: The zone id
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditZoneDto'
      responses:
        '201':
          description: Zone updated
          content:
            application/json:
              schema:
                type: string
                example: OK
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: User Zone not found. Error thrown when user zone is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: User Zone not found
                  error:
                    type: string
                    example: ZONE_NOT_FOUND
      tags:
        - Zone
  /v1/zone/remove/{zoneId}:
    delete:
      operationId: ZoneController_deleteZone
      summary: Delete Zone
      description: User deletes a zone belonging to "zoneId". User must have the "canDelete" permission on zone.
      parameters:
        - name: zoneId
          required: true
          in: path
          description: The zone id
          schema: {}
      responses:
        '200':
          description: Zone deleted
          content:
            application/json:
              schema:
                type: string
                example: OK
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: User Zone not found. Error thrown when user zone is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: User Zone not found
                  error:
                    type: string
                    example: ZONE_NOT_FOUND
      tags:
        - Zone
  /v1/zone/search:
    get:
      operationId: ZoneController_searchZone
      summary: Search Zone
      description: Search zone with requested search term field.
      parameters:
        - name: limit
          required: false
          in: query
          description: The number of records to get.
          schema:
            default: 30
            type: number
        - name: skip
          required: false
          in: query
          description: The number of records to skip.
          schema:
            default: 0
            type: number
        - name: searchTerm
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Search zone.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchZoneResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
      tags:
        - Zone
  /v1/zone/users/list/{zoneId}:
    get:
      operationId: ZoneController_zoneUserList
      summary: List Zone Users
      description: List users of the zone belonging to "zoneId".
      parameters:
        - name: limit
          required: false
          in: query
          description: The number of records to get.
          schema:
            default: 30
            type: number
        - name: skip
          required: false
          in: query
          description: The number of records to skip.
          schema:
            default: 0
            type: number
        - name: name
          required: false
          in: query
          schema:
            type: string
        - name: zoneId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Lists zone users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneUserResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: User Zone not found. Error thrown when user zone is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: User Zone not found
                  error:
                    type: string
                    example: ZONE_NOT_FOUND
      tags:
        - Zone
  /v1/zone/join/{zoneId}:
    post:
      operationId: ZoneController_joinPublicZone
      summary: Join Zone
      description: Current authenticated user joins a zone belonging to "zoneId". The id of the user zone is returned
      parameters:
        - name: zoneId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Joined zone
          content:
            application/json:
              schema:
                type: integer
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: Error thrown when the zone is not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: Zone not found
                  error:
                    type: string
                    example: ZONE_NOT_FOUND
      tags:
        - Zone
  /v1/zone/invite/{zoneId}:
    post:
      operationId: ZoneController_inviteToJoinZone
      summary: Invite User to Zone
      description: Current authenticated user invites a user to specified zone. The id of the invitation is returned. User must have "canInvite" permission on zone permissions.
      parameters:
        - name: zoneId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteToJoinDto'
      responses:
        '201':
          description: Zone user invited
          content:
            application/json:
              schema:
                type: integer
        '400':
          description: Error thrown when the user is already invited
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: The user with the email '<email>' has already been invited to this zone
                  error:
                    type: string
                    example: USER_ALREADY_INVITED_TO_ZONE
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: User Zone not found. Error thrown when user zone is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: User Zone not found
                  error:
                    type: string
                    example: ZONE_NOT_FOUND
        '409':
          description: Error thrown when the user is already member of this zone
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 409
                  message:
                    type: string
                    example: The user with the email '<email>' has already a member of this zone
                  error:
                    type: string
                    example: USER_ALREADY_MEMBER_OF_ZONE
      tags:
        - Zone
  /v1/zone/invitation/response:
    post:
      operationId: ZoneController_respondToInvitationToJoinZone
      summary: Respond Invitation of Zone
      description: Respond invitation of specified zone with "accept" or "reject" enum parameters.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationResponseDto'
      responses:
        '201':
          description: User Responds to invitation
          content:
            application/json:
              schema:
                type: string
                example: OK
        '400':
          description: Error thrown when validation error occurs according to requested payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
        '404':
          description: Error thrown when the requested invitation is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: Invitation not found
                  error:
                    type: string
                    example: INVITATION_NOT_FOUND
      tags:
        - Zone
...
  /v1/zone/display-photo/{fileName}:
    get:
      operationId: ZoneController_viewProfilePhoto
      summary: Get Display Photo
      description: Get display photo stream of zone with "fileName" parameter
      parameters:
        - name: fileName
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Get display photo of selected zone
      tags:
        - Zone
...
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateZoneDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        subdomain:
          type: string
          minimum: 3
          maximum: 32
          pattern: /^(?![0-9]+$)(?!.*-$)(?!-)[a-zA-Z0-9-]{3,32}$/g
          description: 'Subdomain which will be used as zone address. See: https://datatracker.ietf.org/doc/html/rfc1035'
        public:
          type: boolean
      required:
        - name
        - subdomain
    EditZoneDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        subdomain:
          type: string
        public:
          type: boolean
...
    SearchZoneResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SearchZone'
        total:
          type: number
        limit:
          type: number
        skip:
          type: number
      required:
        - data
        - total
        - limit
        - skip
...
    ZoneUserResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ZoneUserItem'
        total:
          type: number
        limit:
          type: number
        skip:
          type: number
      required:
        - data
        - total
        - limit
        - skip
    InviteToJoinDto:
      type: object
      properties:
        email:
          type: string
      required:
        - email
...